+++ src/Series.cpp
@@ -68,17 +68,6 @@ std::string cleanFilename(std::string const& filename, Format f);
 std::function< std::tuple< bool, int >(std::string const&) >
 matcher(std::string const& prefix, int padding, std::string const& postfix, Format f);
 
-struct Series::ParsedInput
-{
-    std::string path;
-    std::string name;
-    Format format;
-    IterationEncoding iterationEncoding;
-    std::string filenamePrefix;
-    std::string filenamePostfix;
-    int filenamePadding;
-};  //ParsedInput
-
 #if openPMD_HAVE_MPI
 Series::Series(std::string const& filepath,
                AccessType at,
@@ -100,6 +89,10 @@ Series::Series(std::string const& filepath,
     auto input = parseInput(filepath);
     auto handler = createIOHandler(input->path, at, input->format);
     init(handler, std::move(input));
+}
+
+Series::Series()
+{
 }
 
 Series::~Series()
@@ -361,10 +354,10 @@ Series::flush()
     IOHandler->flush();
 }
 
-std::unique_ptr< Series::ParsedInput >
+std::unique_ptr< auxiliary::ParsedInput >
 Series::parseInput(std::string filepath)
 {
-    std::unique_ptr< Series::ParsedInput > input{new Series::ParsedInput};
+    std::unique_ptr< auxiliary::ParsedInput > input{new auxiliary::ParsedInput};
 
 #ifdef _WIN32
     if( auxiliary::contains(filepath, '/') )
@@ -462,7 +455,7 @@ Series::parseInput(std::string filepath)
 
 void
 Series::init(std::shared_ptr< AbstractIOHandler > ioHandler,
-             std::unique_ptr< Series::ParsedInput > input)
+             std::unique_ptr< auxiliary::ParsedInput > input)
 {
     m_writable->IOHandler = ioHandler;
     IOHandler = m_writable->IOHandler.get();